pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "focus-groups"
version = "0.1.0"
description = "AI-powered synthetic focus group platform"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "sentence-transformers>=3.0",
    "pgvector>=0.3",
    "numpy>=1.24",
    "anthropic>=0.40",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.scripts]
focus-group = "focus_groups.cli_runner:main"

[tool.setuptools.packages.find]
where = ["src"]
